Pipeline refactoring - move classes to separate files, reorganize directories, add domain for asset sids
